refactor: extract shared TabbedPreviewBlock base class
PlantUMLBlock and MermaidBlock share an identical tabbed image/code
preview skeleton. Extract it into a reusable base class and move
MermaidBlock out of the 5600-line markdown_with_highlight.dart.

Base class (tabbed_preview_block.dart) provides:
- Shell: Container + border + clip + margin
- Header: segmented Image/Code tab control
- Built-in Copy button + extra-actions spacing
- AnimatedSwitcher body with previewHeight = 406
- Shared code view (dual-axis scroll + SelectableText)
- PreviewTabButton / PreviewTextAction / PreviewBlockColors
- PreviewLoadingView / PreviewErrorView

PlantUMLBlock:      464 -> 83 lines (-381)
MermaidBlock:       extracted to own file (712 lines)
markdown_with_highlight: 5606 -> ~4650 lines (-957)
Mermaid loading/error views preserved as private subclass overrides.
