+++ lib/shared/widgets/markdown_with_highlight.dart
@@ -23,9 +23,8 @@ import '../../features/chat/pages/image_viewer_page.dart';
 import '../../features/chat/pages/html_preview_page.dart';
 import 'snackbar.dart';
 import 'ios_tactile.dart';
-import 'mermaid_bridge.dart';
 import 'export_capture_scope.dart';
-import 'mermaid_image_cache.dart';
+import 'mermaid_block.dart';
 import 'plantuml_block.dart';
 import 'package:path/path.dart' as p;
 import 'package:Kelivo/l10n/app_localizations.dart';
@@ -479,7 +478,7 @@ class _MarkdownWithCodeHighlightState extends State<MarkdownWithCodeHighlight> {
         final lang = name.trim();
         final restoredCode = _unmaskHtmlTagStartsInsideFencedCode(code);
         if (lang.toLowerCase() == 'mermaid') {
-          return _MermaidBlock(
+          return MermaidBlock(
             code: restoredCode,
             streaming: widget.streaming && !closed,
           );
@@ -2468,53 +2467,6 @@ String _codeBlockStateKey(String language, String code) {
   return '$normalizedLanguage|$anchor';
 }
 
-String _mermaidCacheKey(
-  String code,
-  bool isDark,
-  Map<String, String> themeVars,
-) {
-  final entries = themeVars.entries.toList()
-    ..sort((a, b) => a.key.compareTo(b.key));
-  final themeSig = entries.map((e) => '${e.key}=${e.value}').join('&');
-  return '${isDark ? 'dark' : 'light'}|$themeSig|$code';
-}
-
-enum MermaidBitmapRenderStatus { success, failed, unsupported }
-
-class MermaidBitmapRenderResult {
-  const MermaidBitmapRenderResult._(this.status, [this.bytes]);
-
-  factory MermaidBitmapRenderResult.success(Uint8List bytes) {
-    return MermaidBitmapRenderResult._(
-      MermaidBitmapRenderStatus.success,
-      bytes,
-    );
-  }
-
-  factory MermaidBitmapRenderResult.failed() {
-    return const MermaidBitmapRenderResult._(MermaidBitmapRenderStatus.failed);
-  }
-
-  factory MermaidBitmapRenderResult.unsupported() {
-    return const MermaidBitmapRenderResult._(
-      MermaidBitmapRenderStatus.unsupported,
-    );
-  }
-
-  final MermaidBitmapRenderStatus status;
-  final Uint8List? bytes;
-}
-
-typedef MermaidBitmapRenderOverride =
-    Future<MermaidBitmapRenderResult> Function(
-      String code,
-      bool isDark,
-      Map<String, String> themeVars,
-    );
-
-@visibleForTesting
-MermaidBitmapRenderOverride? debugMermaidBitmapRenderOverride;
-
 class _CodeBlockHeaderToggle extends StatelessWidget {
   const _CodeBlockHeaderToggle({
     required this.expanded,
@@ -3495,912 +3447,6 @@ String _csvCell(String value) {
   return '"${value.replaceAll('"', '""')}"';
 }
 
-class _MermaidBlock extends StatefulWidget {
-  final String code;
-  final bool streaming;
-  const _MermaidBlock({required this.code, required this.streaming});
-
-  @override
-  State<_MermaidBlock> createState() => _MermaidBlockState();
-}
-
-enum _MermaidTab { image, code }
-
-class _MermaidBlockState extends State<_MermaidBlock> {
-  static const Duration _streamingBitmapRenderDelay = Duration(
-    milliseconds: 360,
-  );
-  static const Duration _settledBitmapRenderDelay = Duration(milliseconds: 220);
-  static const double _previewHeight = 406;
-
-  _MermaidTab _selectedTab = _MermaidTab.image;
-  late final ScrollController _vMermaidScrollController;
-  OverlayEntry? _renderOverlayEntry;
-  bool _renderQueued = false;
-  bool _renderingBitmap = false;
-  String? _renderKey;
-  Uint8List? _lastRenderedBytes;
-  Timer? _streamingRenderDebounce;
-  bool _bitmapRenderingUnsupported = false;
-  bool _suppressBitmapLoading = false;
-  final Set<String> _failedBitmapRenderKeys = <String>{};
-
-  @override
-  Widget build(BuildContext context) {
-    final cs = Theme.of(context).colorScheme;
-    final isDark = Theme.of(context).brightness == Brightness.dark;
-    final l10n = AppLocalizations.of(context)!;
-
-    final mermaidColors = _MermaidBlockColors.resolve(isDark);
-
-    // Build theme variables mapping for Mermaid from Material ColorScheme
-    String hex(Color c) {
-      final v = c.toARGB32();
-      final r = (v >> 16) & 0xFF;
-      final g = (v >> 8) & 0xFF;
-      final b = v & 0xFF;
-      return '#'
-              '${r.toRadixString(16).padLeft(2, '0')}'
-              '${g.toRadixString(16).padLeft(2, '0')}'
-              '${b.toRadixString(16).padLeft(2, '0')}'
-          .toUpperCase();
-    }
-
-    final themeVars = <String, String>{
-      'primaryColor': hex(cs.primary),
-      'primaryTextColor': hex(cs.onPrimary),
-      'primaryBorderColor': hex(cs.primary),
-      'secondaryColor': hex(cs.secondary),
-      'secondaryTextColor': hex(cs.onSecondary),
-      'secondaryBorderColor': hex(cs.secondary),
-      'tertiaryColor': hex(cs.tertiary),
-      'tertiaryTextColor': hex(cs.onTertiary),
-      'tertiaryBorderColor': hex(cs.tertiary),
-      'background': hex(cs.surface),
-      'mainBkg': hex(cs.primaryContainer),
-      'secondBkg': hex(cs.secondaryContainer),
-      'lineColor': hex(cs.onSurface),
-      'textColor': hex(cs.onSurface),
-      'nodeBkg': hex(cs.surface),
-      'nodeBorder': hex(cs.primary),
-      'clusterBkg': hex(cs.surface),
-      'clusterBorder': hex(cs.primary),
-      'actorBorder': hex(cs.primary),
-      'actorBkg': hex(cs.surface),
-      'actorTextColor': hex(cs.onSurface),
-      'actorLineColor': hex(cs.primary),
-      'taskBorderColor': hex(cs.primary),
-      'taskBkgColor': hex(cs.primary),
-      'taskTextLightColor': hex(cs.onPrimary),
-      'taskTextDarkColor': hex(cs.onSurface),
-      'labelColor': hex(cs.onSurface),
-      'errorBkgColor': hex(cs.error),
-      'errorTextColor': hex(cs.onError),
-    };
-
-    final exporting = ExportCaptureScope.of(context);
-    final cacheKey = _mermaidCacheKey(widget.code, isDark, themeVars);
-    final themedCachedBytes = MermaidImageCache.get(cacheKey);
-    final legacyCachedBytes = MermaidImageCache.get(widget.code);
-    final prefixCachedBytes = widget.streaming
-        ? _findCachedStreamingMermaidPrefix(
-            widget.code,
-            isDark: isDark,
-            themeVars: themeVars,
-          )
-        : null;
-    final exactCachedBytes = themedCachedBytes ?? legacyCachedBytes;
-    final cachedBytes = exactCachedBytes ?? prefixCachedBytes;
-    final displayBytes =
-        cachedBytes ?? (widget.streaming ? _lastRenderedBytes : null);
-    final actionBytes = cachedBytes ?? displayBytes;
-    final renderFailedForCurrentCode = _failedBitmapRenderKeys.contains(
-      cacheKey,
-    );
-    final hasRenderableCode = widget.code.trim().isNotEmpty;
-    if (!exporting &&
-        hasRenderableCode &&
-        exactCachedBytes == null &&
-        !_bitmapRenderingUnsupported &&
-        !renderFailedForCurrentCode) {
-      _scheduleBitmapRender(
-        isDark: isDark,
-        themeVars: themeVars,
-        delay: widget.streaming
-            ? _streamingBitmapRenderDelay
-            : _settledBitmapRenderDelay,
-      );
-    }
-    final hasImage = displayBytes != null && displayBytes.isNotEmpty;
-    final showLoading =
-        !hasImage &&
-        !_suppressBitmapLoading &&
-        !_bitmapRenderingUnsupported &&
-        !renderFailedForCurrentCode &&
-        (widget.streaming || _renderQueued || _renderingBitmap);
-    final showError =
-        !hasImage &&
-        !_bitmapRenderingUnsupported &&
-        renderFailedForCurrentCode &&
-        _selectedTab == _MermaidTab.image;
-
-    return Container(
-      width: double.infinity,
-      margin: const EdgeInsets.symmetric(vertical: 6),
-      decoration: BoxDecoration(
-        color: mermaidColors.body,
-        borderRadius: BorderRadius.circular(8),
-      ),
-      clipBehavior: Clip.antiAlias,
-      foregroundDecoration: BoxDecoration(
-        borderRadius: BorderRadius.circular(8),
-        border: Border.all(color: mermaidColors.border),
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Container(
-            decoration: BoxDecoration(
-              color: mermaidColors.header,
-              border: Border(
-                bottom: BorderSide(color: mermaidColors.border, width: 1),
-              ),
-            ),
-            padding: const EdgeInsets.symmetric(vertical: 5),
-            child: Row(
-              children: [
-                Expanded(
-                  child: Padding(
-                    padding: const EdgeInsetsDirectional.only(
-                      start: 16,
-                      end: 10,
-                    ),
-                    child: Align(
-                      alignment: AlignmentDirectional.centerStart,
-                      child: DecoratedBox(
-                        decoration: BoxDecoration(
-                          color: mermaidColors.tabTrack,
-                          borderRadius: BorderRadius.circular(8),
-                        ),
-                        child: Padding(
-                          padding: const EdgeInsets.all(2),
-                          child: Row(
-                            mainAxisSize: MainAxisSize.min,
-                            children: [
-                              _MermaidTabButton(
-                                label: l10n.mermaidImageTab,
-                                selected: _selectedTab == _MermaidTab.image,
-                                colors: mermaidColors,
-                                onTap: () {
-                                  setState(
-                                    () => _selectedTab = _MermaidTab.image,
-                                  );
-                                },
-                              ),
-                              _MermaidTabButton(
-                                label: l10n.mermaidCodeTab,
-                                selected: _selectedTab == _MermaidTab.code,
-                                colors: mermaidColors,
-                                onTap: () {
-                                  setState(
-                                    () => _selectedTab = _MermaidTab.code,
-                                  );
-                                },
-                              ),
-                            ],
-                          ),
-                        ),
-                      ),
-                    ),
-                  ),
-                ),
-                if (!exporting)
-                  Padding(
-                    padding: const EdgeInsetsDirectional.only(end: 10),
-                    child: Row(
-                      mainAxisSize: MainAxisSize.min,
-                      children: [
-                        _MermaidTextAction(
-                          icon: Lucide.Copy,
-                          label: l10n.shareProviderSheetCopyButton,
-                          colors: mermaidColors,
-                          onTap: () => _copyMermaidCode(context),
-                        ),
-                        const SizedBox(width: 4),
-                        _MermaidTextAction(
-                          icon: Lucide.Download,
-                          label: l10n.mermaidExportPng,
-                          colors: mermaidColors,
-                          enabled:
-                              actionBytes != null && actionBytes.isNotEmpty,
-                          onTap: actionBytes == null || actionBytes.isEmpty
-                              ? null
-                              : () => _saveMermaidBytes(context, actionBytes),
-                        ),
-                        const SizedBox(width: 4),
-                        _MermaidTextAction(
-                          icon: Lucide.Maximize2,
-                          label: l10n.mermaidFullScreen,
-                          colors: mermaidColors,
-                          enabled:
-                              actionBytes != null && actionBytes.isNotEmpty,
-                          onTap: actionBytes == null || actionBytes.isEmpty
-                              ? null
-                              : () => _openMermaidImageViewer(
-                                  context,
-                                  actionBytes,
-                                ),
-                        ),
-                      ],
-                    ),
-                  ),
-              ],
-            ),
-          ),
-          SizedBox(
-            key: const ValueKey('mermaid-preview-body'),
-            width: double.infinity,
-            height: _previewHeight,
-            child: ColoredBox(
-              color: mermaidColors.body,
-              child: AnimatedSwitcher(
-                duration: const Duration(milliseconds: 180),
-                switchInCurve: Curves.easeOutCubic,
-                switchOutCurve: Curves.easeInCubic,
-                layoutBuilder: (currentChild, previousChildren) {
-                  return currentChild ?? const SizedBox.shrink();
-                },
-                child: _buildMermaidBody(
-                  context: context,
-                  isDark: isDark,
-                  colors: mermaidColors,
-                  displayBytes: displayBytes,
-                  cacheKey: cacheKey,
-                  showLoading: showLoading,
-                  showError: showError,
-                ),
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildMermaidBody({
-    required BuildContext context,
-    required bool isDark,
-    required _MermaidBlockColors colors,
-    required Uint8List? displayBytes,
-    required String cacheKey,
-    required bool showLoading,
-    required bool showError,
-  }) {
-    if (_selectedTab == _MermaidTab.code ||
-        _bitmapRenderingUnsupported ||
-        widget.code.trim().isEmpty) {
-      return _buildMermaidCodeView(context, isDark);
-    }
-
-    if (displayBytes != null && displayBytes.isNotEmpty) {
-      return Padding(
-        key: ValueKey<String>('mermaid-image-$cacheKey'),
-        padding: const EdgeInsets.all(8),
-        child: MouseRegion(
-          cursor: SystemMouseCursors.click,
-          child: GestureDetector(
-            behavior: HitTestBehavior.opaque,
-            onTap: () => _openMermaidImageViewer(context, displayBytes),
-            child: Image(image: MemoryImage(displayBytes), fit: BoxFit.contain),
-          ),
-        ),
-      );
-    }
-
-    if (showLoading) {
-      return _MermaidLoadingView(
-        key: const ValueKey('mermaid-loading-body'),
-        colors: colors,
-      );
-    }
-
-    if (showError) {
-      return _MermaidErrorView(
-        key: const ValueKey('mermaid-error-body'),
-        colors: colors,
-      );
-    }
-
-    return _buildMermaidCodeView(context, isDark);
-  }
-
-  Widget _buildMermaidCodeView(BuildContext context, bool isDark) {
-    final codeView = SelectableHighlightView(
-      widget.code,
-      language: 'plaintext',
-      theme: _transparentBgTheme(
-        isDark ? atomOneDarkReasonableTheme : githubTheme,
-      ),
-      padding: EdgeInsets.zero,
-      textStyle: TextStyle(fontFamily: 'monospace', fontSize: 13, height: 1.5),
-    );
-
-    return Padding(
-      key: const ValueKey('mermaid-code-body'),
-      padding: const EdgeInsets.all(12),
-      child: ScrollConfiguration(
-        behavior: ScrollConfiguration.of(context).copyWith(
-          dragDevices: {
-            ui.PointerDeviceKind.touch,
-            ui.PointerDeviceKind.mouse,
-            ui.PointerDeviceKind.stylus,
-            ui.PointerDeviceKind.unknown,
-          },
-        ),
-        child: Scrollbar(
-          controller: _vMermaidScrollController,
-          thumbVisibility: true,
-          interactive: true,
-          notificationPredicate: (notif) => notif.metrics.axis == Axis.vertical,
-          child: SingleChildScrollView(
-            controller: _vMermaidScrollController,
-            child: SingleChildScrollView(
-              scrollDirection: Axis.horizontal,
-              child: codeView,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    _vMermaidScrollController = ScrollController();
-  }
-
-  @override
-  void didUpdateWidget(covariant _MermaidBlock oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    if (oldWidget.code != widget.code ||
-        oldWidget.streaming != widget.streaming) {
-      _suppressBitmapLoading = false;
-      if (!widget.streaming || oldWidget.streaming != widget.streaming) {
-        _streamingRenderDebounce?.cancel();
-        _renderQueued = false;
-        _renderingBitmap = false;
-        _removeRenderOverlay();
-        _renderKey = null;
-      }
-    }
-    if (widget.code.trim().isEmpty) {
-      _lastRenderedBytes = null;
-      _suppressBitmapLoading = false;
-      _bitmapRenderingUnsupported = false;
-      _failedBitmapRenderKeys.clear();
-    }
-  }
-
-  @override
-  void dispose() {
-    _streamingRenderDebounce?.cancel();
-    _removeRenderOverlay();
-    _vMermaidScrollController.dispose();
-    super.dispose();
-  }
-
-  void _scheduleBitmapRender({
-    required bool isDark,
-    required Map<String, String> themeVars,
-    required Duration delay,
-  }) {
-    if (_renderQueued || _renderingBitmap) return;
-    _renderQueued = true;
-    _streamingRenderDebounce?.cancel();
-    _streamingRenderDebounce = Timer(delay, () {
-      _renderQueued = false;
-      if (!mounted) return;
-      _renderBitmap(isDark: isDark, themeVars: themeVars);
-    });
-  }
-
-  Future<void> _renderBitmap({
-    required bool isDark,
-    required Map<String, String> themeVars,
-  }) async {
-    final code = widget.code;
-    final cacheKey = _mermaidCacheKey(code, isDark, themeVars);
-    if (MermaidImageCache.get(cacheKey) != null) return;
-    final renderOverride = debugMermaidBitmapRenderOverride;
-    final overlay = renderOverride == null ? Overlay.maybeOf(context) : null;
-    if (renderOverride == null && overlay == null) {
-      _markBitmapRenderingUnsupported(cacheKey);
-      return;
-    }
-    setState(() {
-      _renderKey = cacheKey;
-      _renderingBitmap = true;
-    });
-
-    MermaidBitmapRenderResult result = MermaidBitmapRenderResult.failed();
-    try {
-      result = renderOverride == null
-          ? await _renderMermaidBitmapWithOverlay(
-              overlay!,
-              code,
-              isDark,
-              themeVars,
-            )
-          : await renderOverride(code, isDark, themeVars);
-      if (!mounted || _renderKey != cacheKey) return;
-      final bytes = result.bytes;
-      if (result.status == MermaidBitmapRenderStatus.success &&
-          bytes != null &&
-          bytes.isNotEmpty) {
-        MermaidImageCache.put(cacheKey, bytes);
-        _failedBitmapRenderKeys.remove(cacheKey);
-      }
-    } catch (e, st) {
-      debugPrint('Mermaid bitmap render failed: $e\n$st');
-    } finally {
-      if (mounted && _renderKey == cacheKey) {
-        _removeRenderOverlay();
-        setState(() {
-          if (result.status == MermaidBitmapRenderStatus.success &&
-              result.bytes != null &&
-              result.bytes!.isNotEmpty) {
-            _lastRenderedBytes = result.bytes;
-          } else if (result.status == MermaidBitmapRenderStatus.unsupported) {
-            _bitmapRenderingUnsupported = true;
-            _suppressBitmapLoading = true;
-          } else {
-            _failedBitmapRenderKeys.add(cacheKey);
-            _suppressBitmapLoading = true;
-          }
-          _renderingBitmap = false;
-        });
-      }
-    }
-  }
-
-  Future<MermaidBitmapRenderResult> _renderMermaidBitmapWithOverlay(
-    OverlayState overlay,
-    String code,
-    bool isDark,
-    Map<String, String> themeVars,
-  ) async {
-    _removeRenderOverlay();
-    final renderKey = GlobalKey();
-    final handle = createMermaidView(
-      code,
-      isDark,
-      themeVars: themeVars,
-      viewKey: renderKey,
-    );
-    if (handle == null) return MermaidBitmapRenderResult.unsupported();
-
-    _renderOverlayEntry = OverlayEntry(
-      builder: (context) => Positioned(
-        left: -10000,
-        top: -10000,
-        child: ConstrainedBox(
-          constraints: const BoxConstraints.tightFor(width: 720, height: 600),
-          child: Material(color: Colors.transparent, child: handle.widget),
-        ),
-      ),
-    );
-    overlay.insert(_renderOverlayEntry!);
-
-    return _captureMermaidBitmap(handle);
-  }
-
-  void _markBitmapRenderingUnsupported(String cacheKey) {
-    if (!mounted) return;
-    _streamingRenderDebounce?.cancel();
-    _removeRenderOverlay();
-    setState(() {
-      if (_renderKey == null || _renderKey == cacheKey) {
-        _renderKey = null;
-        _renderQueued = false;
-        _renderingBitmap = false;
-      }
-      _bitmapRenderingUnsupported = true;
-    });
-  }
-
-  Uint8List? _findCachedStreamingMermaidPrefix(
-    String code, {
-    required bool isDark,
-    required Map<String, String> themeVars,
-  }) {
-    final lines = code.split('\n');
-    for (var end = lines.length - 1; end >= 1; end--) {
-      final candidate = lines.take(end).join('\n').trimRight();
-      if (candidate.isEmpty) continue;
-      final themed = MermaidImageCache.get(
-        _mermaidCacheKey(candidate, isDark, themeVars),
-      );
-      final legacy = MermaidImageCache.get(candidate);
-      final bytes = themed ?? legacy;
-      if (bytes != null && bytes.isNotEmpty) {
-        _lastRenderedBytes = bytes;
-        return bytes;
-      }
-    }
-    return null;
-  }
-
-  Future<MermaidBitmapRenderResult> _captureMermaidBitmap(
-    MermaidViewHandle handle,
-  ) async {
-    final exportBytes = handle.exportPngBytes;
-    if (exportBytes == null) return MermaidBitmapRenderResult.unsupported();
-    await WidgetsBinding.instance.endOfFrame;
-    await Future<void>.delayed(const Duration(milliseconds: 120));
-    for (var i = 0; i < 4; i++) {
-      try {
-        final bytes = await exportBytes().timeout(
-          const Duration(milliseconds: 900),
-          onTimeout: () => null,
-        );
-        if (bytes != null && bytes.isNotEmpty) {
-          return MermaidBitmapRenderResult.success(bytes);
-        }
-      } catch (e) {
-        if (e is UnsupportedError) {
-          return MermaidBitmapRenderResult.unsupported();
-        }
-        // Mermaid/WebView can report readiness before pixel capture is available.
-      }
-      await Future<void>.delayed(const Duration(milliseconds: 120));
-    }
-    return MermaidBitmapRenderResult.failed();
-  }
-
-  Future<void> _copyMermaidCode(BuildContext context) async {
-    final copiedMessage = AppLocalizations.of(
-      context,
-    )!.chatMessageWidgetCopiedToClipboard;
-    await Clipboard.setData(ClipboardData(text: widget.code));
-    if (!context.mounted) return;
-    showAppSnackBar(
-      context,
-      message: copiedMessage,
-      type: NotificationType.success,
-    );
-  }
-
-  Future<void> _saveMermaidBytes(BuildContext context, Uint8List bytes) async {
-    final l10n = AppLocalizations.of(context)!;
-    final ok = await _saveCachedMermaidPng(bytes);
-    if (!context.mounted) return;
-    if (!ok) {
-      showAppSnackBar(
-        context,
-        message: l10n.mermaidExportFailed,
-        type: NotificationType.error,
-      );
-    } else if (Platform.isAndroid || Platform.isIOS) {
-      showAppSnackBar(
-        context,
-        message: l10n.imageViewerPageSaveSuccess,
-        type: NotificationType.success,
-      );
-    }
-  }
-
-  void _openMermaidImageViewer(BuildContext context, Uint8List bytes) {
-    final src = 'data:image/png;base64,${base64Encode(bytes)}';
-    final provider = MemoryImage(bytes);
-    Navigator.of(context).push(
-      PageRouteBuilder(
-        pageBuilder: (_, __, ___) =>
-            ImageViewerPage(images: [src], imageProviders: {src: provider}),
-        transitionDuration: const Duration(milliseconds: 300),
-        reverseTransitionDuration: const Duration(milliseconds: 240),
-        transitionsBuilder: (context, anim, sec, child) {
-          final curved = CurvedAnimation(
-            parent: anim,
-            curve: Curves.easeOutCubic,
-            reverseCurve: Curves.easeInCubic,
-          );
-          return FadeTransition(opacity: curved, child: child);
-        },
-      ),
-    );
-  }
-
-  void _removeRenderOverlay() {
-    try {
-      _renderOverlayEntry?.remove();
-    } catch (_) {}
-    _renderOverlayEntry = null;
-  }
-
-  Future<bool> _saveCachedMermaidPng(Uint8List bytes) async {
-    try {
-      final l10n = AppLocalizations.of(context)!;
-      final suggested = 'mermaid_${DateTime.now().millisecondsSinceEpoch}.png';
-      if (Platform.isWindows || Platform.isLinux || Platform.isMacOS) {
-        final savePath = await FilePicker.platform.saveFile(
-          dialogTitle: l10n.backupPageExportToFile,
-          fileName: suggested,
-          type: FileType.custom,
-          allowedExtensions: const ['png'],
-        );
-        if (savePath == null || savePath.isEmpty) return false;
-        await File(savePath).parent.create(recursive: true);
-        await File(savePath).writeAsBytes(bytes);
-        return true;
-      }
-      final result = await ImageGallerySaverPlus.saveImage(
-        bytes,
-        quality: 100,
-        name: 'kelivo-mermaid-${DateTime.now().millisecondsSinceEpoch}',
-      );
-      if (result is Map) {
-        final isSuccess =
-            result['isSuccess'] == true || result['isSuccess'] == 1;
-        final filePath = result['filePath'] ?? result['file_path'];
-        return isSuccess || (filePath is String && filePath.isNotEmpty);
-      }
-    } catch (_) {}
-    return false;
-  }
-}
-
-class _MermaidBlockColors {
-  const _MermaidBlockColors({
-    required this.body,
-    required this.header,
-    required this.border,
-    required this.tabTrack,
-    required this.tabSelected,
-    required this.textPrimary,
-    required this.textSecondary,
-    required this.textTertiary,
-  });
-
-  final Color body;
-  final Color header;
-  final Color border;
-  final Color tabTrack;
-  final Color tabSelected;
-  final Color textPrimary;
-  final Color textSecondary;
-  final Color textTertiary;
-
-  static _MermaidBlockColors resolve(bool isDark) {
-    if (isDark) {
-      return const _MermaidBlockColors(
-        body: Color(0xFF212121),
-        header: Color(0xFF303030),
-        border: Color(0xFF383838),
-        tabTrack: Color(0xF2212121),
-        tabSelected: Color(0xFF333333),
-        textPrimary: Color(0xFFE6E6E6),
-        textSecondary: Color(0xFFA0A0A0),
-        textTertiary: Color(0xFF707070),
-      );
-    }
-
-    return const _MermaidBlockColors(
-      body: Color(0xFFF8F8F8),
-      header: Color(0xFFEDEDED),
-      border: Color(0xFFE0E0E0),
-      tabTrack: Color(0xCCD9D9D9),
-      tabSelected: Color(0xFFFFFFFF),
-      textPrimary: Color(0xFF261208),
-      textSecondary: Color(0xFF46352B),
-      textTertiary: Color(0xFF5B4C43),
-    );
-  }
-}
-
-class _MermaidTabButton extends StatefulWidget {
-  const _MermaidTabButton({
-    required this.label,
-    required this.selected,
-    required this.colors,
-    required this.onTap,
-  });
-
-  final String label;
-  final bool selected;
-  final _MermaidBlockColors colors;
-  final VoidCallback onTap;
-
-  @override
-  State<_MermaidTabButton> createState() => _MermaidTabButtonState();
-}
-
-class _MermaidTabButtonState extends State<_MermaidTabButton> {
-  bool _pressed = false;
-  bool _hovered = false;
-
-  @override
-  Widget build(BuildContext context) {
-    final baseColor = widget.selected
-        ? widget.colors.tabSelected
-        : Colors.transparent;
-    final hoverColor = Color.alphaBlend(
-      widget.colors.textPrimary.withValues(alpha: _pressed ? 0.10 : 0.06),
-      baseColor,
-    );
-    final bg = widget.selected || _pressed || _hovered
-        ? hoverColor
-        : Colors.transparent;
-
-    return Semantics(
-      button: true,
-      selected: widget.selected,
-      label: widget.label,
-      child: MouseRegion(
-        cursor: SystemMouseCursors.click,
-        onEnter: (_) => setState(() => _hovered = true),
-        onExit: (_) => setState(() {
-          _hovered = false;
-          _pressed = false;
-        }),
-        child: GestureDetector(
-          behavior: HitTestBehavior.opaque,
-          onTapDown: (_) => setState(() => _pressed = true),
-          onTapCancel: () => setState(() => _pressed = false),
-          onTapUp: (_) => setState(() => _pressed = false),
-          onTap: widget.onTap,
-          child: AnimatedContainer(
-            duration: const Duration(milliseconds: 160),
-            curve: Curves.easeOutCubic,
-            padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 2),
-            decoration: BoxDecoration(
-              color: bg,
-              borderRadius: BorderRadius.circular(6),
-            ),
-            child: SelectionContainer.disabled(
-              child: Text(
-                widget.label,
-                style: TextStyle(
-                  fontSize: 13,
-                  height: 1.35,
-                  fontWeight: widget.selected
-                      ? AppFontWeights.semibold
-                      : AppFontWeights.medium,
-                  color: widget.selected
-                      ? widget.colors.textPrimary
-                      : widget.colors.textSecondary,
-                ),
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _MermaidTextAction extends StatelessWidget {
-  const _MermaidTextAction({
-    required this.icon,
-    required this.label,
-    required this.colors,
-    this.onTap,
-    this.enabled = true,
-  });
-
-  final IconData icon;
-  final String label;
-  final _MermaidBlockColors colors;
-  final VoidCallback? onTap;
-  final bool enabled;
-
-  @override
-  Widget build(BuildContext context) {
-    final active = enabled && onTap != null;
-    final color = colors.textSecondary.withValues(alpha: active ? 0.88 : 0.38);
-
-    return Tooltip(
-      message: label,
-      child: IosIconButton(
-        onTap: onTap,
-        enabled: active,
-        semanticLabel: label,
-        color: color,
-        padding: const EdgeInsets.symmetric(horizontal: 6, vertical: 5),
-        builder: (buttonColor) => Icon(icon, size: 14, color: buttonColor),
-      ),
-    );
-  }
-}
-
-class _MermaidLoadingView extends StatefulWidget {
-  const _MermaidLoadingView({super.key, required this.colors});
-
-  final _MermaidBlockColors colors;
-
-  @override
-  State<_MermaidLoadingView> createState() => _MermaidLoadingViewState();
-}
-
-class _MermaidLoadingViewState extends State<_MermaidLoadingView>
-    with SingleTickerProviderStateMixin {
-  late final AnimationController _controller;
-
-  @override
-  void initState() {
-    super.initState();
-    _controller = AnimationController(
-      vsync: this,
-      duration: const Duration(milliseconds: 1100),
-    )..repeat();
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = AppLocalizations.of(context)!;
-    return Center(
-      child: Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          RotationTransition(
-            turns: _controller,
-            child: Icon(
-              Lucide.Loader,
-              size: 24,
-              color: widget.colors.textSecondary,
-            ),
-          ),
-          const SizedBox(width: 8),
-          Text(
-            l10n.mermaidGeneratingImage,
-            style: TextStyle(
-              fontSize: 14,
-              height: 1.3,
-              color: widget.colors.textSecondary,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class _MermaidErrorView extends StatelessWidget {
-  const _MermaidErrorView({super.key, required this.colors});
-
-  final _MermaidBlockColors colors;
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = AppLocalizations.of(context)!;
-    return Center(
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Icon(Lucide.ImageOff, size: 48, color: colors.textTertiary),
-          const SizedBox(height: 8),
-          Text(
-            l10n.mermaidGenerationFailedHint,
-            textAlign: TextAlign.center,
-            style: TextStyle(
-              fontSize: 14,
-              height: 1.35,
-              color: colors.textTertiary,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
 // Full-width horizontal rule with softer color
 class SoftHrLine extends BlockMd {
   @override
@@ -4455,7 +3501,7 @@ class FencedCodeBlockMd extends BlockMd {
     final langLower = lang.toLowerCase();
     final isStreamingFence = streaming && !closed;
     if (langLower == 'mermaid') {
-      return _MermaidBlock(code: code, streaming: isStreamingFence);
+      return MermaidBlock(code: code, streaming: isStreamingFence);
     } else if (langLower == 'plantuml') {
       return PlantUMLBlock(code: code);
     }
